requirements.txt
asgiref==3.6.0
Django==4.1.7
django-admin==2.0.2
django-excel-response2==3.0.5
django-filter==22.1
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
excel-base==1.0.4
factory-boy==3.2.1
Faker==17.6.0
isoweek==1.3.3
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
TimeConvert==3.0.13
tzdata==2022.7
tzlocal==4.2
xlwt==1.3.0
